feat:: add some logic

+++ backend-server/src/main/java/com/example/backend_server/services/OutfitService.java
@@ -1,5 +1,6 @@
 package com.example.backend_server.services;
 
+import com.example.backend_server.exceptions.OutfitNotFoundException;
 import com.example.backend_server.models.dto.OutfitDto;
 import com.example.backend_server.models.entitys.Outfit;
 import com.example.backend_server.repositories.JPAOutfitRepository;
@@ -8,10 +9,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @Service
 @Transactional
 public class OutfitService {
-    private JPAOutfitRepository repository;
+    private final JPAOutfitRepository repository;
 
     @Autowired
     public OutfitService(JPAOutfitRepository repository) {this.repository = repository;}
@@ -20,9 +23,15 @@ public class OutfitService {
         return new OutfitDto(repository.save(new Outfit(dto)));
     }
 
+    // update тут еще свои проверки
+
     public OutfitDto getById(long id) {
-        // ошибка если не существует
-        return new OutfitDto(repository.getOne(id));
+        Optional<Outfit> toFind = repository.findById(id);
+        if (toFind.isPresent()) {
+            Outfit outfit = toFind.get();
+            return new OutfitDto(outfit);
+        }
+        throw new OutfitNotFoundException("outfit with id " + id + "didn't exist");
     }
 
     public OutfitDto getByEntity(OutfitDto dto) {
@@ -30,8 +39,12 @@ public class OutfitService {
     }
 
     public void deleteById(long id) {
-        // проверка на существование
-        repository.deleteById(id);
+        Optional<Outfit> toFind = repository.findById(id);
+        if (toFind.isPresent()) {
+            repository.deleteById(id);
+            return;
+        }
+        throw new OutfitNotFoundException("outfit with id " + id + "didn't exist");
     }
 
     public void deleteByEntity(OutfitDto dto) {
